Using Gates and Can to implement authoriization

+++ app/Http/Controllers/QuestionsController.php
@@ -65,7 +65,12 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Access Denied");
+        }
         return view("questions.edit", compact("question"));
+
+
     }
 
     /**
@@ -77,6 +82,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Access Denied");
+        }
         $question->update($request->only('title', 'body'));
         return redirect()->route('questions.index')->with('success', "Your question has been updated");
     }
@@ -89,6 +97,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)) {
+            abort(403, "Access Denied");
+        }
         $question->delete();
         return redirect('/questions')->with('success', 'Your question has been removed');
     }
